Move replay creation logic to ScoreDatabase.
